added Read, Write and Add ParticleLoop access for CellDat
